chore: move CompleteExchangeError into ExchangeSdkError and delete SwapError
SwapError.ts had been reduced to just CompleteExchangeError and
CompleteExchangeStep after the error hierarchy consolidation. Move
those into ExchangeSdkError.ts and delete the file.
