--- lib/src/error/SwapError.ts
@@ -1,19 +1,0 @@
-export type CompleteExchangeStep =
-  | "INIT"
-  | "SET_PARTNER_KEY"
-  | "CHECK_PARTNER"
-  | "PROCESS_TRANSACTION"
-  | "CHECK_TRANSACTION_SIGNATURE"
-  | "CHECK_PAYOUT_ADDRESS"
-  | "CHECK_REFUND_ADDRESS"
-  | "SIGN_COIN_TRANSACTION";
-
-export class CompleteExchangeError extends Error {
-  step: CompleteExchangeStep;
-
-  constructor(step: CompleteExchangeStep, message?: string) {
-    super(message);
-    this.name = "CompleteExchangeError";
-    this.step = step;
-  }
-}
